backend/requirements.txt
asttokens==2.4.1
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
executing==2.1.0
Flask==3.0.3
Flask-Cors==5.0.0
icecream==2.1.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pip==22.0.2
Pygments==2.18.0
pymongo==4.10.1
PyYAML==6.0.2
requests==2.32.3
setuptools==59.6.0
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.2
cryptography==44.0.1
